+++ src/Charts/Gauge/Gauge.js
