+++ src/Charts/Gauge/index.js
